% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwlist.R, R/bwlistDoc.R
\name{addURLBlackList}
\alias{addURLBlackList}
\title{Set a URL to blacklist}
\usage{
addURLBlackList(proxy, regex, statusCode, method = NULL, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{regex}{The blacklist regular expression.}

\item{statusCode}{The HTTP status code to return for URLs that are 
blacklisted.}

\item{method}{The regular expression for matching HTTP method 
(GET, POST, PUT, etc). Optional, by default processing all HTTP 
method.}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Sets a URL to blacklist
}
\examples{
\dontrun{
  
}
}
\seealso{
Other bwlist functions: \code{\link{clearBlackList}},
  \code{\link{clearWhiteList}}, \code{\link{getBlackList}},
  \code{\link{getWhiteList}}, \code{\link{setWhiteList}}
}

