% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmproxy.R, R/bmpserver.R
\docType{package}
\name{bmproxy}
\alias{bmproxy}
\alias{bmproxy-package}
\title{Wrapper around the 'BrowserMob' Proxy REST API for R}
\usage{
bmproxy(port = 9090L, version = "latest", proxyPortRange = 39500L:39999L)
}
\arguments{
\item{port}{Port to run on}

\item{version}{what version of chromedriver to run. Default = "latest"
which runs the most recent version. To see other version currently
sourced run binman::list_versions("chromedriver")}

\item{proxyPortRange}{The range of ports to use for proxies as integers.}
}
\value{
Returns a list with named elements process, output, error
    and stop. process is the output from calling \code{\link{spawn_process}}
    output, error and stop are functions calling
    \code{\link{process_read}}, \code{\link{process_read}} with "stderr"
    pipe and \code{\link{process_kill}}  respectively  on process.
}
\description{
Wrapper around the 'BrowserMob Proxy' REST API for R.
'BrowserMob Proxy' is a utility to help web developers watch and
manipulate network traffic from their AJAX applications.

Start a Browser Mob proxy
}
\examples{
\dontrun{
bmPxy <- bmproxy()
bmPxy$stop()
}
}

