% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwlist.R, R/bwlistDoc.R
\name{clearBlackList}
\alias{clearBlackList}
\title{Clear Black List}
\usage{
clearBlackList(proxy, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Clears the black list of URL's
}
\examples{
\dontrun{
  
}
}
\seealso{
Other bwlist functions: \code{\link{addURLBlackList}},
  \code{\link{clearWhiteList}}, \code{\link{getBlackList}},
  \code{\link{getWhiteList}}, \code{\link{setWhiteList}}
}

