% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R, R/sessionDoc.R
\name{emptyDNS}
\alias{emptyDNS}
\title{Empty the DNS}
\usage{
emptyDNS(proxy, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Empties the DNS cache. Clears the DNS cache associated with the proxy instance
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
}
}
\seealso{
Other session functions: \code{\link{closeProxy}},
  \code{\link{getPorts}}, \code{\link{openPort}},
  \code{\link{setRetry}}, \code{\link{timeouts}},
  \code{\link{waitForTraffic}}
}

