% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwlist.R, R/bwlistDoc.R
\name{getBlackList}
\alias{getBlackList}
\title{getBlackList}
\usage{
getBlackList(proxy, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{...}{Additonal function arguments}
}
\value{
Returns a list. Each item of the list is a regex from the
}
\description{
getBlackList
}
\examples{
\dontrun{
  
}
}
\seealso{
Other bwlist functions: \code{\link{addURLBlackList}},
  \code{\link{clearBlackList}},
  \code{\link{clearWhiteList}}, \code{\link{getWhiteList}},
  \code{\link{setWhiteList}}
}

