% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R, R/sessionDoc.R
\name{getPorts}
\alias{getPorts}
\title{Get ports/proxies}
\usage{
getPorts(proxy, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{...}{Additonal function arguments}
}
\value{
Returns a list of ports. Each port is a \code{\link{proxy}}
   object
}
\description{
returns the ports that are currently open on the Browser Mob Proxy 
   server.
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
}
}
\seealso{
Other session functions: \code{\link{closeProxy}},
  \code{\link{emptyDNS}}, \code{\link{openPort}},
  \code{\link{setRetry}}, \code{\link{timeouts}},
  \code{\link{waitForTraffic}}
}

