% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwlist.R, R/bwlistDoc.R
\name{getWhiteList}
\alias{getWhiteList}
\title{Get white list}
\usage{
getWhiteList(proxy, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{...}{Additonal function arguments}
}
\value{
Returns a list. Each item of the list is a regex from the 
   white list
}
\description{
Gets a list of URL patterns from the whitelist
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
}
}
\seealso{
Other bwlist functions: \code{\link{addURLBlackList}},
  \code{\link{clearBlackList}},
  \code{\link{clearWhiteList}}, \code{\link{getBlackList}},
  \code{\link{setWhiteList}}
}

