% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{httr_proxy}
\alias{httr_proxy}
\title{Set httr proxy configuration}
\usage{
httr_proxy(proxy, cert_file = NULL)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{cert_file}{An SSL certification file. If NULL bmproxy will use
the  certificate file bundled with Browser Mob Proxy. For details
see https://github.com/lightbody/browsermob-proxy/blob/master/mitm/README.md}
}
\value{
Returns a \code{\link{httr}} config object see 
   \code{\link{config}}. This config contains the url and port of the 
   proxy together with the appropriate certificate file.
}
\description{
Sets the proxy configuration for use with \code{\link{httr}}
}
\examples{
\dontrun{
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
  # create a HAR
  prxy \%>\% createHAR(ref = "httr_traffic")
  rproj <- GET("https://www.r-project.org/", httr_proxy(prxy))
  httr_har <- prxy \%>\% getHAR()
  httr_har[["log"]][["entries"]][[1]]["request"]
  httr_har[["log"]][["entries"]][[1]]["response"]
  prxy \%>\% closeProxy
}
}

