% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R, R/sessionDoc.R
\name{openPort}
\alias{openPort}
\title{Open a new port/proxy}
\usage{
openPort(proxy, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{...}{Additonal function arguments}
}
\description{
opens a new port on the BrowserMob Proxy server
This function is usually called via \code{\link{proxy}}
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
  # get ports with running proxies
  prxy \%>\% getPorts
  # set timeouts
  prxy \%>\% timeouts(requestTimeout = 50000L, readTimeout = 50000L,
                    connectionTimeout = 50000L)
  # close proxy and check ports
  prxy \%>\% 
    closeProxy \%>\% 
    getPorts
}
}
\seealso{
Other session functions: \code{\link{closeProxy}},
  \code{\link{emptyDNS}}, \code{\link{getPorts}},
  \code{\link{setRetry}}, \code{\link{timeouts}},
  \code{\link{waitForTraffic}}
}

