% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite.R, R/overwriteDoc.R
\name{remapHosts}
\alias{remapHosts}
\title{Remap hosts}
\usage{
remapHosts(proxy, address = NULL, ipAddress = NULL, hostMap = list(), ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{address}{url that you wish to remap}

\item{ipAddress}{IP Address that will handle all traffic for the 
address passed in}

\item{hostMap}{An optional named list of address/ipAddress items.}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Remap the hosts for a specific URL. Option to add a list of address
   remaps via hostMap argument.
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
}
}
\seealso{
Other overwrite functions: \code{\link{removeRewrites}},
  \code{\link{rewriteURL}}, \code{\link{setAutoBAuth}},
  \code{\link{setHeaders}}
}

