% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite.R, R/overwriteDoc.R
\name{removeRewrites}
\alias{removeRewrites}
\title{Remove redirection rules}
\usage{
removeRewrites(proxy, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Removes all URL redirection rules currently in effect
}
\examples{
\dontrun{
  
}
}
\seealso{
Other overwrite functions: \code{\link{remapHosts}},
  \code{\link{rewriteURL}}, \code{\link{setAutoBAuth}},
  \code{\link{setHeaders}}
}

