% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercept.R, R/interceptDoc.R
\name{resInterceptor}
\alias{resInterceptor}
\title{Intercept response}
\usage{
resInterceptor(proxy, jsFilter, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{jsFilter}{Javascript response filters have access to the 
variables response (type io.netty.handler.codec.http.HttpResponse), 
contents (type net.lightbody.bmp.util.HttpMessageContents), and 
messageInfo (type net.lightbody.bmp.util.HttpMessageInfo). As in the 
request filter, messageInfo contains additional information about 
the message.}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Executes the javascript against each response
}
\examples{
\dontrun{
  
}
}
\seealso{
Other intercept functions: \code{\link{reqInterceptor}}
}

