% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite.R, R/overwriteDoc.R
\name{rewriteURL}
\alias{rewriteURL}
\title{Rewrite URL}
\usage{
rewriteURL(proxy, match, replace, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{match}{a regex to match requests with}

\item{replace}{a string to replace the matches with}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Rewrites the requested url.
}
\examples{
\dontrun{
  
}
}
\seealso{
Other overwrite functions: \code{\link{remapHosts}},
  \code{\link{removeRewrites}}, \code{\link{setAutoBAuth}},
  \code{\link{setHeaders}}
}

