% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sel_proxy}
\alias{sel_proxy}
\title{Set Selenium proxy configuration}
\usage{
sel_proxy(proxy, eCaps = NULL, browser = c("firefox", "chrome"),
  bmpIPaddress = NULL)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{eCaps}{A list of existing extra capabilities.}

\item{browser}{The browser type to set the config for. Can use firefox 
or chrome. The default is firefox. If left NULL firefox default is 
used.}
}
\value{
Returns an extra capabilities list that can be passed to 
   Selenium

bmpIPaddress Stipulate an alternative BMP ip address. The
   Selenium server may for example be running in a docker container
   as may the BMP server. Defaults to NULL and the ip address implied
   by proxy is used
}
\description{
Sets the proxy configuration for use with Selenium
}
\examples{
\dontrun{
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
  eCap <- sel_proxy(prxy, browser = "chrome")
}
}

