% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite.R, R/overwriteDoc.R
\name{setAutoBAuth}
\alias{setAutoBAuth}
\title{Set basic authentication}
\usage{
setAutoBAuth(proxy, domain, username, password, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{domain}{domain to set authentication credentials for}

\item{username}{valid username to use when authenticating}

\item{password}{valid password to use when authenticating}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Adds automatic basic authentication to a domain
}
\examples{
\dontrun{
  
}
}
\seealso{
Other overwrite functions: \code{\link{remapHosts}},
  \code{\link{removeRewrites}}, \code{\link{rewriteURL}},
  \code{\link{setHeaders}}
}

