% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite.R, R/overwriteDoc.R
\name{setHeaders}
\alias{setHeaders}
\title{Set headers}
\usage{
setHeaders(proxy, headers = list(), ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{headers}{A named list. The name of an item is the header name 
(such as "User-Agent") and value is a value of HTTP header to setup 
(such as "BrowserMob-Agent"). 
Example: list("User-Agent" = "BrowserMob-Agent")}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Set and override HTTP Request headers. This sets the headers that will 
   set by the proxy on all requests
}
\examples{
\dontrun{
  library(httr)
  library(xml2)
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
  prxy \%>\% createHAR(ref = "httr_traffic_1")
  ua1 <- GET("http://httpbin.org/user-agent", httr_proxy(prxy), 
             user_agent("httr"))
  
  prxy \%>\% 
    setHeaders(list("User-Agent" = "BrowserMob-Agent")) \%>\% 
    newPage(ref = "httr_traffic_2")
  ua2 <- GET("http://httpbin.org/user-agent", httr_proxy(prxy), 
             user_agent("httr"))
  
  prxy \%>\% remapHosts("www.r-project.org", "www.google.com")
  roogle <- GET("http://www.r-project.org", httr_proxy(prxy))
  roogle \%>\% read_html \%>\% xml_find_all("//a") \%>\% xml_attrs
  
  prxy \%>\% 
    newPage(ref = "httr_traffic_3")
  auth1 <-GET("http://httpbin.org/basic-auth/user/passwd", 
              httr_proxy(prxy))
  prxy \%>\% 
    setAutoBAuth("httpbin.org",
                 "user", "passwd")
  auth2 <-GET("http://httpbin.org/basic-auth/user/passwd", 
              httr_proxy(prxy))
  
  auth1$status_code
  auth2$status_code
  content(auth2)
  
  prxy \%>\% closeProxy
  
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
  prxy  \%>\% 
    rewriteURL(match = 'http://www\\\\.google\\\\.com.*', 
               replace = 'http://www\\\\.r-project\\\\.org')
  gproj <- GET("http://www.google.com/ncr", httr_proxy(prxy))
  gproj \%>\% read_html \%>\% xml_find_all("//title") \%>\% xml_text

  prxy \%>\% removeRewrites
  gproj <- GET("http://www.google.com/ncr", httr_proxy(prxy))
  gproj \%>\% read_html \%>\% xml_find_all("//title") \%>\% xml_text
  
  prxy \%>\% closeProxy
}
}
\seealso{
Other overwrite functions: \code{\link{remapHosts}},
  \code{\link{removeRewrites}}, \code{\link{rewriteURL}},
  \code{\link{setAutoBAuth}}
}

