% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwlist.R, R/bwlistDoc.R
\name{setWhiteList}
\alias{setWhiteList}
\title{Set white list}
\usage{
setWhiteList(proxy, regex, statusCode, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{regex}{a character vector of regular expressions}

\item{statusCode}{the HTTP status code to return for URLs nthat do not 
match the whitelist}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Sets a list of URL patterns to whitelist
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
}
}
\seealso{
Other bwlist functions: \code{\link{addURLBlackList}},
  \code{\link{clearBlackList}},
  \code{\link{clearWhiteList}}, \code{\link{getBlackList}},
  \code{\link{getWhiteList}}
}

