% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R, R/sessionDoc.R
\name{timeouts}
\alias{timeouts}
\title{timeouts}
\usage{
timeouts(proxy, requestTimeout = -1L, readTimeout = 60000L,
  connectionTimeout = 60000L, dnsCacheTimeout = 0L, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{requestTimeout}{Request timeout in milliseconds. A timeout value 
of -1 is interpreted as infinite timeout. Optional, default to -1}

\item{readTimeout}{ead timeout in milliseconds. Which is the timeout 
for waiting for data or, put differently, a maximum period 
inactivity between two consecutive data packets). A timeout value of 
zero is interpreted as an infinite timeout. Optional, default to 
60000.}

\item{connectionTimeout}{Determines the timeout in milliseconds until a 
connection is established. A timeout value of zero is interpreted as 
an infinite timeout. Optional, default to 60000.}

\item{dnsCacheTimeout}{Sets the maximum length of time that records 
will be stored in this Cache. A nonpositive value disables this 
feature (that is, sets no limit). Optional, default to "0".}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
timeouts
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
  # get ports with running proxies
  prxy \%>\% getPorts
  # set timeouts
  prxy \%>\% timeouts(requestTimeout = 50000L, readTimeout = 50000L,
                    connectionTimeout = 50000L)
  # close proxy and check ports
  prxy \%>\% 
    closeProxy \%>\% 
    getPorts
}
}
\seealso{
Other session functions: \code{\link{closeProxy}},
  \code{\link{emptyDNS}}, \code{\link{getPorts}},
  \code{\link{openPort}}, \code{\link{setRetry}},
  \code{\link{waitForTraffic}}
}

