% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R, R/sessionDoc.R
\name{waitForTraffic}
\alias{waitForTraffic}
\title{Wait for traffic to stop}
\usage{
waitForTraffic(proxy, quietPeriod = NULL, timeOut = NULL, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{quietPeriod}{number of milliseconds the network needs to be quiet 
for}

\item{timeOut}{max number of milliseconds to wait}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Waits for the network to be quiet
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
}
}
\seealso{
Other session functions: \code{\link{closeProxy}},
  \code{\link{emptyDNS}}, \code{\link{getPorts}},
  \code{\link{openPort}}, \code{\link{setRetry}},
  \code{\link{timeouts}}
}

