% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwlist.R, R/bwlistDoc.R
\name{addURLBlackList}
\alias{addURLBlackList}
\title{Set a URL to blacklist}
\usage{
addURLBlackList(proxy, regex, statusCode, method = NULL, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{regex}{The blacklist regular expression.}

\item{statusCode}{The HTTP status code to return for URLs that are 
blacklisted.}

\item{method}{The regular expression for matching HTTP method 
(GET, POST, PUT, etc). Optional, by default processing all HTTP 
method.}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Sets a URL to blacklist
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
  library(httr)
  # use httr_proxy to configure proxy for httr
  white_list <- prxy \%>\% getWhiteList
  white_list
  # set an entry in the white list
  prxy \%>\% setWhiteList(".*r-project\\\\.org.*", statusCode = 333L)
  white_list <- prxy \%>\% getWhiteList
  white_list
  # expect passing sites
  rcran <- GET("https://cran.r-project.org/", httr_proxy(prxy))
  rproj <- GET("https://www.r-project.org/", httr_proxy(prxy))
  # expect 333 status code for non matching url
  goog <- GET("https://www.google.com/", httr_proxy(prxy))
  goog$status_code
  
  # clear the white list
  white_list <- prxy \%>\% 
    clearWhiteList \%>\% 
    getWhiteList
  length(white_list)
  
  # get the black list
  black_list <- prxy \%>\% getBlackList
  black_list
  
  # add site to black list
  prxy \%>\% addURLBlackList(".*r-project\\\\.org.*", statusCode = 334L)
  goog <- GET("https://www.google.com/", httr_proxy(prxy))
  
  # expect sites to return 334 status code
  rcran <- GET("https://cran.r-project.org/", httr_proxy(prxy))
  rproj <- GET("https://www.r-project.org/", httr_proxy(prxy))
  rcran$status_code
  rproj$status_code
  
  # clear black list
  black_list <- prxy \%>\% 
    clearBlackList \%>\% 
    getBlackList
  length(black_list)

  prxy \%>\% closeProxy
}
}
\seealso{
Other bwlist functions: \code{\link{clearBlackList}},
  \code{\link{clearWhiteList}}, \code{\link{getBlackList}},
  \code{\link{getWhiteList}}, \code{\link{setWhiteList}}
}

