% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmproxy.R
\docType{package}
\name{bmproxy}
\alias{bmproxy}
\alias{bmproxy-package}
\title{Wrapper around the 'BrowserMob' Proxy REST API for R}
\description{
Wrapper around the 'BrowserMob Proxy' REST API for R.
'BrowserMob Proxy' is a utility to help web developers watch and
manipulate network traffic from their AJAX applications.
}

