% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwlist.R, R/bwlistDoc.R
\name{getBlackList}
\alias{getBlackList}
\title{Get BlackList}
\usage{
getBlackList(proxy, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{...}{Additonal function arguments}
}
\value{
Returns a data.frame of black listed urls
}
\description{
Gets the URL's in the Black List. Returns a data.frame
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
}
}
\seealso{
Other bwlist functions: \code{\link{addURLBlackList}},
  \code{\link{clearBlackList}},
  \code{\link{clearWhiteList}}, \code{\link{getWhiteList}},
  \code{\link{setWhiteList}}
}

