% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R, R/bandwidthDoc.R
\name{getBWremaining}
\alias{getBWremaining}
\title{Get bandwidth remaining}
\usage{
getBWremaining(proxy, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{...}{Additonal function arguments}
}
\value{
A data.frame detailing the maximum bandwidth and remaining
   bandwidth upstream and downstream is returned.
}
\description{
Displays the amount of data remaining to be uploaded/downloaded until 
   the limit is reached
}
\examples{
\dontrun{
  
}
}
\seealso{
Other bandwidth functions: \code{\link{limitBandwidth}}
}

