% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har.R, R/harDoc.R
\name{getHAR}
\alias{getHAR}
\title{Get HAR}
\usage{
getHAR(proxy, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{...}{Additonal function arguments}
}
\value{
Returns the HAR as parsed JSON content
}
\description{
Returns the JSON/HAR content representing all the HTTP traffic passed 
   through the proxy (provided you have already created the HAR with 
   \code{\link{createHAR}})
}
\examples{
\dontrun{
  
}
}
\seealso{
Other har functions: \code{\link{createHAR}},
  \code{\link{newPage}}
}

