% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R, R/bandwidthDoc.R
\name{limitBandwidth}
\alias{limitBandwidth}
\title{Limit Bandwidth}
\usage{
limitBandwidth(proxy, downK = NULL, upK = NULL, downMaxK = NULL,
  upMaxK = NULL, latency = NULL, enable = FALSE, payPercent = NULL,
  maxBits = NULL, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{downK}{Sets the downstream bandwidth limit in kbps. Optional if 
NULL (default) unlimited.}

\item{upK}{Sets the upstream bandwidth limit kbps. Optional if 
NULL (default) unlimited.}

\item{downMaxK}{Specifies how many kilobytes in total the client is 
allowed to download through the proxy. Optional if NULL (default) 
unlimited.}

\item{upMaxK}{Specifies how many kilobytes in total the client is 
allowed to upload through the proxy. Optional if NULL (default) 
unlimited.}

\item{latency}{Add the given latency to each HTTP request. Optional, by 
default all requests are invoked without latency.}

\item{enable}{A boolean that enable bandwidth limiter. Optional, by 
default to FALSE, but setting any of the bandwidth arguments will 
implicitly enable throttling}

\item{payPercent}{Specifying what percentage of data sent is payload, 
e.g. use this to take into account overhead due to tcp/ip. Optional.}

\item{maxBits}{The max bits per seconds you want this instance of 
StreamManager to respect. Optional.}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Limit the bandwidth through the proxy on the port associated with 
   proxy argument.
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
}
}
\seealso{
Other bandwidth functions: \code{\link{getBWremaining}}
}

