% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har.R, R/harDoc.R
\name{newPage}
\alias{newPage}
\title{Create new page}
\usage{
newPage(proxy, ref = NULL, title = ref, ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{ref}{The string name of the first page ref that should be used in 
the HAR. Optional, default to "Page N" where N is the next page 
number (zero indexed so 1st page is page 0).}

\item{title}{The title of new HAR page. Optional, default to ref.}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Starts a new page on the existing HAR.
}
\examples{
\dontrun{
  # A BMP server is assummed running on port 9090
  prxy <- proxy(bmpPort = 9090L, port = 39500L)
  library(httr)
  prxy \%>\% createHAR(ref = "httr_traffic_1")
  # use httr_proxy to configure proxy for httr
  rproj <- GET("https://www.r-project.org/", httr_proxy(prxy))
  httr_har <- prxy \%>\% getHAR()
  httr_har[["log"]][["entries"]][[1]]["request"]
  httr_har[["log"]][["entries"]][[1]]["response"]
  
  # set new page
  prxy \%>\% newPage(ref = "httr_traffic_2")
  rcran <- GET("https://cran.r-project.org/", httr_proxy(prxy))
  httr_har <- prxy \%>\% getHAR()
  vapply(httr_har[["log"]][["pages"]], '[[', character(1), "id")
  # [1] "httr_traffic_1" "httr_traffic_2"
  prxy \%>\% closeProxy
}
}
\seealso{
Other har functions: \code{\link{createHAR}},
  \code{\link{getHAR}}
}

