% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite.R, R/overwriteDoc.R
\name{setHeaders}
\alias{setHeaders}
\title{Set headers}
\usage{
setHeaders(proxy, headers = list(), ...)
}
\arguments{
\item{proxy}{An object of class "proxy". A proxy object see
\code{\link{proxy}}.}

\item{headers}{A named list. The name of an item is the header name 
(such as "User-Agent") and value is a value of HTTP header to setup 
(such as "BrowserMob-Agent"). 
Example: list("User-Agent" = "BrowserMob-Agent")}

\item{...}{Additonal function arguments}
}
\value{
invisible(proxy): An object of class "proxy" is invisibly
   returned. A proxy object see \code{\link{proxy}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
Set and override HTTP Request headers. This sets the headers that will 
   set by the proxy on all requests
}
\examples{
\dontrun{
  
}
}
\seealso{
Other overwrite functions: \code{\link{remapHosts}},
  \code{\link{removeRewrites}}, \code{\link{rewriteURL}},
  \code{\link{setAutoBAuth}}
}

